Player Pawn Setup
Began working on setting up the player pawn. This somehow led to creating movement controls. The current setup does not fix the issue of the ship not being invisible and the current movement controls do not move the ship nor the camera. This issue requires more research to finish but I'm too tired to keep going for tonight.