UI System Rework Part 1
Started working on a new UI system. Created UI_Master as the super class for UI_Manager and UI_Controller. Created UI_Controllers for the main menu, multiplayer menu, and options menu. Created a UI_Manager for the main menu. Buttons on the titleUI function correctly. Other menus have not been implemented yet. Escape key functionality was added but the logical code needs to be changed to support future UIs. Perhaps moving the logic to the UI_Master class is needed?