Refactor 97% Complete. Fixed UI bugs.
Finished refactoring the codebase outside the hotbar. Fixed bugs with the ui where the item slots were not registering properly, the inventory slot widget was not connected and thus always failed to construct, the ui managers defaulted to titleUI, and some UIs had improper variables. Added a c++ class for the hud and made the ui manager a proper referencable variable.