Reworked Mouse Slot. Fixed Visual Issues With Inventory UI.
Added a mouse slot to the player controller and the inventory ui. The two prior objects and all inventories now link to the same mouse slot via pointers. This means they can now share the same mouse item! Old code was slightly changed but mostly left alone since it still works. A sprite now follows the mouse! Although it needs to be moved relative to the mouse still. In the process, an issue with the inventory ui clipping and not appearing in the correct size has been fixed.