Movement Rework. Reworked Vertical Movement.
Changed vertical movement to be component based. Changed movement to be target location based. Made some changes to replication and desync detection. Movement on the server will now move at the same rate as the client instead of moving immediately towards the latest location. Added an array that tracks positions and vectors. Game is currently in a broken state.